% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pf_core.R
\name{load_pf_core}
\alias{load_pf_core}
\title{Helper to load core PartyFacts data}
\usage{
load_pf_core(countries = NULL)
}
\arguments{
\item{countries}{A character vector of VDEM country codes.}
}
\value{
A tibble with PartyFacts data for the countries of interest.
}
\description{
This function allows to conveniently load PartyFacts data and subset the countries of interest.
}
\examples{

library(partymakeR)

\dontrun{

# Define countries of interest
countries <- c("AFG", "AIA")

# Load data
dat_pf_core <- load_pf_core(countries = countries)

}

}
