% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pf_wide.R
\name{load_pf_external}
\alias{load_pf_external}
\title{Helper to load wide version of PartyFacts external data}
\usage{
load_pf_external(countries = NULL, keep_sources = NULL)
}
\arguments{
\item{countries}{A character vector of VDEM country codes.}

\item{keep_sources}{A character vector of dataset names included in PartyFacts external data. These include:
"manifesto", "parlgov", "wikipedia", "ches", "clea", "ejpr", "vparty", "chisols", "dpi", "essprtc", "essprtv", "huber", "kitschelt", "polcon", "ppmd", "whogov", "wvs", "afrelec", "afro", "ccdd", "ccs", "coppedge", "cses", "elecglob", "epac", "hix", "erdda", "euandi", "gloelec", "gpd", "gps", "ipod", "janda", "jw", "kurep", "laeda", "latino", "laverhunt", "leadglob", "mackie", "mapp", "morgan", "mudde", "nped", "parlspeech", "postyug", "pip", "poppa", "ppdb", "ppla", "ppmdall", "ray", "tap", "populist", "voteview", "ees14", "ppepe", "thomas"

Note: source data are matched with a \code{\link[base]{regex}}, so "manif" works just like "manifesto", and "manifesto|parlg" returns both data from Manifesto and ParlGov datasets.}
}
\value{
A tibble with PartyFacts data for the countries and datasets of interest.
}
\description{
Helper to load wide version of PartyFacts external data
}
\examples{

library(partymakeR)

\dontrun{

# Define countries of interest
countries <- c("AFG", "AIA")

# Load data
dat_pf_external <- load_pf_external(countries = countries, keep_sources = "manifesto")
dat_pf_external

}

}
