% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_intermediate.R
\name{see_intermediate}
\alias{see_intermediate}
\title{A helper to check content of /data_intermediate directory}
\usage{
see_intermediate(... = NULL)
}
\arguments{
\item{...}{A set of quoted nested folder names.}
}
\value{
A tibble with a column of file names and a column of relative paths.
}
\description{
This function assumes a standardized RProject folder structure with a "data_intermediate" folder at top level. The function is a wrapper of \pkg{here} and allows to quickly inspect files and folders.
}
\examples{

library(partymakeR)

\dontrun{

# See main data_input directory
see_intermediate()

# See top level
see_intermediate("..")

# Enter any number of nested folders
see_intermediate("folder", "subfolder")

}

}
